+++ care/emr/resources/medication/dispense/spec.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from enum import Enum
 
 from pydantic import UUID4, BaseModel, model_validator
@@ -123,8 +124,8 @@ class MedicationDispenseWriteSpec(BaseMedicationDispenseSpec):
     location: UUID4
     authorizing_request: UUID4 | None = None
     item: UUID4
-    quantity: float
-    days_supply: float | None = None
+    quantity: Decimal
+    days_supply: Decimal | None = None
     fully_dispensed: bool | None = None
     order: UUID4 | None = None
     create_dispense_order: CreateDispenseOrder | None = None
@@ -206,7 +207,7 @@ class MedicationDispenseReadSpec(BaseMedicationDispenseSpec):
     created_date: datetime
     modified_date: datetime
     location: dict
-    quantity: float
+    quantity: Decimal
     authorizing_request: dict | None = None
     order: dict | None = None
